update files has added

+++ helpers/fileUtils.js
@@ -1,8 +1,9 @@
 import { v4 as uuidv4 } from 'uuid';
 import * as url from 'url';
 import path from 'path'
-import { LOCAL_PREVIOUS_PATH, LOCAL_UPLOAD_PATH } from '../constant/routes.constant.js';
+import { LOCAL_ASSETS_FOLDER_PATH, LOCAL_PREVIOUS_PATH, LOCAL_UPLOAD_PATH } from '../constant/routes.constant.js';
 import fs from 'fs';
+import { DEFAULT_IMAGE } from '../constant/values.constant.js';
 
 export const __dirname = url.fileURLToPath(new URL(LOCAL_PREVIOUS_PATH, import.meta.url));
 
@@ -11,6 +12,16 @@ export const generateNameFile = (file) => `${uuidv4()}.${getExtFromFile(file)}`
 export const getExtFromFile = (file) => {
     const cutName = file.name.split('.');
     return cutName[cutName.length - 1];
+}
+
+export const getFile = (nameFile = '', folder = '') => {
+    const imagePath = path.join(__dirname, LOCAL_UPLOAD_PATH, folder, nameFile);
+
+    if (nameFile !== '' && fs.existsSync(imagePath)) {
+        return imagePath;
+    }
+
+    return path.join(__dirname, LOCAL_ASSETS_FOLDER_PATH, '', DEFAULT_IMAGE);
 }
 
 export const uploadFile = (file, folder = '') => {
@@ -29,10 +40,10 @@ export const uploadFile = (file, folder = '') => {
     })
 }
 
-export const deleteFile = (nameFile, folder = '') => {
+export const deleteFile = (nameFile = '', folder = '') => {
     const imagePath = path.join(__dirname, LOCAL_UPLOAD_PATH, folder, nameFile);
 
-    if (fs.existsSync(imagePath)) {
+    if (nameFile !== '' && fs.existsSync(imagePath)) {
         fs.unlinkSync(imagePath)
     }
 }
